LayoutTests/platform/win/animations/animation-time-expected.txt

Tests that window.webkitAnimationTime remains the same during a script run and that all animations started within a script run are synchronized

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL startAnimationTime should be defined. Was undefined
PASS startAnimationTime is endAnimationTime
FAIL window.getComputedStyle(div1).left should be 42px. Was 53px.
PASS successfullyParsed is true

TEST COMPLETE
